Ticket-003: minor refactoring
