+++ src/main/java/de/maxi_bauer/board/GameMove.java
@@ -1,4 +1,4 @@
-package de.maxi_bauer.data;
+package de.maxi_bauer.board;
 
 public record GameMove(int row, int column) {
 }
