+++ src/main/java/de/maxi_bauer/player/Player.java
@@ -1,6 +1,6 @@
 package de.maxi_bauer.player;
 
-import de.maxi_bauer.data.GameMove;
+import de.maxi_bauer.board.GameMove;
 
 public interface Player {
     GameMove getMove();
